services/opa-audit-sink/obs/otel_boot.py

import os
from opentelemetry import trace, metrics
from opentelemetry.sdk.resources import Resource
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.metrics import MeterProvider
from opentelemetry.sdk.trace.export import BatchSpanProcessor
from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
from opentelemetry.exporter.otlp.proto.grpc.metric_exporter import OTLPMetricExporter

service_name = os.getenv("OTEL_SERVICE_NAME", "opa-audit-sink")
otel_endpoint = os.getenv("OTEL_EXPORTER_OTLP_ENDPOINT", "http://otel-collector.default.svc:4317")

resource = Resource.create({"service.name": service_name})
tp = TracerProvider(resource=resource)
tp.add_span_processor(BatchSpanProcessor(OTLPSpanExporter(endpoint=otel_endpoint, insecure=True)))
trace.set_tracer_provider(tp)

metrics.set_meter_provider(MeterProvider(resource=resource))
# Optional: metrics via OTLP – many collectors prefer pull; keep SDK for future push
